remove nu_cli crate dependency from nu_std
