+++ crates/nu-cli/src/commands.rs
@@ -1,9 +1,9 @@
 use log::info;
 use miette::Result;
-use nu_command::util::report_error;
 use nu_engine::{convert_env_values, eval_block};
 use nu_parser::parse;
 use nu_protocol::engine::Stack;
+use nu_protocol::report_error;
 use nu_protocol::{
     engine::{EngineState, StateWorkingSet},
     PipelineData, Spanned, Value,
