+++ crates/nu-cli/src/config_files.rs
@@ -1,8 +1,8 @@
 use crate::util::eval_source;
-use nu_command::util::report_error;
 #[cfg(feature = "plugin")]
 use nu_path::canonicalize_with;
 use nu_protocol::engine::{EngineState, Stack, StateWorkingSet};
+use nu_protocol::report_error;
 use nu_protocol::{HistoryFileFormat, PipelineData};
 #[cfg(feature = "plugin")]
 use nu_protocol::{ParseError, Spanned};
