+++ crates/nu-cli/src/prompt_update.rs
@@ -1,7 +1,7 @@
 use crate::NushellPrompt;
 use log::trace;
-use nu_command::util::report_error;
 use nu_engine::eval_subexpression;
+use nu_protocol::report_error;
 use nu_protocol::{
     engine::{EngineState, Stack, StateWorkingSet},
     Config, PipelineData, Value,
