+++ crates/nu-protocol/src/cli_error.rs
@@ -1,4 +1,4 @@
-use crate::engine::StateWorkingSet;
+use crate::engine::{EngineState, StateWorkingSet};
 use miette::{LabeledSpan, MietteHandlerOpts, ReportHandler, RgbColors, Severity, SourceCode};
 use thiserror::Error;
 
@@ -16,6 +16,27 @@ pub fn format_error(
     error: &(dyn miette::Diagnostic + Send + Sync + 'static),
 ) -> String {
     format!("Error: {:?}", CliError(error, working_set))
+}
+
+pub fn report_error(
+    working_set: &StateWorkingSet,
+    error: &(dyn miette::Diagnostic + Send + Sync + 'static),
+) {
+    eprintln!("Error: {:?}", CliError(error, working_set));
+    // reset vt processing, aka ansi because illbehaved externals can break it
+    #[cfg(windows)]
+    {
+        let _ = nu_utils::enable_vt_processing();
+    }
+}
+
+pub fn report_error_new(
+    engine_state: &EngineState,
+    error: &(dyn miette::Diagnostic + Send + Sync + 'static),
+) {
+    let working_set = StateWorkingSet::new(engine_state);
+
+    report_error(&working_set, error);
 }
 
 impl std::fmt::Debug for CliError<'_> {
