+++ crates/nu-std/Cargo.toml
@@ -9,6 +9,5 @@ version = "0.78.1"
 
 [dependencies]
 miette = { version = "5.6.0", features = ["fancy-no-backtrace"] }
-nu-cli = { version = "0.78.1", path = "../nu-cli" }
 nu-parser = { version = "0.78.1", path = "../nu-parser" }
 nu-protocol = { version = "0.78.1", path = "../nu-protocol" }
